+++ plugin-manifest.json
@@ -44,8 +44,8 @@
       }
     },
     {
-      "label": "Episode",
-      "value": "episode",
+      "label": "Episodes",
+      "value": "episodes",
       "documentation": {
         "link": "https://github.com/applicaster/zapp-pipes-provider-mpx-new#readme",
         "input_description": "",
@@ -54,8 +54,8 @@
       }
     },
     {
-      "label": "Movie",
-      "value": "movie",
+      "label": "Movies",
+      "value": "movies",
       "documentation": {
         "link": "https://github.com/applicaster/zapp-pipes-provider-mpx-new#readme",
         "input_description": "",
