+++ src/provider/handler/mappers/moviesMapper.js
@@ -1,9 +1,8 @@
-import * as R from 'ramda'
+import * as R from "ramda";
 import {convertDate, createEntry} from "../../../utils";
 import {types} from "../../../types";
 
-
-export function mapEpisode(episode) {
+export function mapMovies(movies) {
 
   const {
     id,
@@ -13,20 +12,20 @@ export function mapEpisode(episode) {
     description,
     credits,
     tags,
-    tvSeasonEpisodeNumber,
-    tvSeasonNumber,
     media: [
       {
-        publicUrl: src
+        publicUrl: src,
+        availableDate: availableAtDate
       }
     ],
     distributionRightIds,
     guid,
     thumbnails: images
-  } = episode;
+  } = movies;
 
   const published = convertDate(publishedAt, 'LL');
   const updated = convertDate(updatedAt);
+  const availableDate = convertDate(availableAtDate);
 
   const genre = R.filter(R.propEq('scheme', 'Genre'))(tags);
 
@@ -45,8 +44,7 @@ export function mapEpisode(episode) {
 
   const extensions = {
     alternate_id: guid,
-    tvSeasonNumber: `Season ${tvSeasonNumber}`,
-    tvSeasonEpisodeNumber: `Episode ${tvSeasonEpisodeNumber}`,
+    availableDate,
     distributionRightIds,
     genre,
     credits,
